langchain/rechnungsbucher/requirements.txt
langchain>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-anthropic>=0.3,<0.4
langgraph>=0.3,<0.4
pydantic>=2.0,<3.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
pymupdf>=1.25,<2.0
lxml>=5.0,<6.0
rich>=13.0,<14.0
